+++ src/wl_utils/wl_conversion.py
@@ -31,14 +31,6 @@ def to_iso_639_3(main, lang_code):
 def to_iso_639_1(main, lang_code):
     return main.settings_global['lang_codes'][lang_code]
 
-def to_text_type_code(main, text_type_text):
-    return main.settings_global['text_types'][text_type_text]
-
-def to_text_type_text(main, text_type_code):
-    for text, code in main.settings_global['text_types'].items():
-        if text_type_code == code:
-            return text
-
 def to_encoding_code(main, encoding_text):
     return main.settings_global['file_encodings'][encoding_text]